+++ packages/api/tests/endpoints/collections/delete-collection.js
@@ -12,11 +12,10 @@ const {
   generateLocalTestDb,
   destroyLocalTestDb,
   CollectionPgModel,
+  migrationDir,
 } = require('@cumulus/db');
 const { bootstrapElasticSearch } = require('@cumulus/es-client/bootstrap');
 const { Search } = require('@cumulus/es-client/search');
-
-const { migrationDir } = require('../../../../../lambdas/db-migration');
 
 const models = require('../../../models');
 const {